:recycle: Split method- and property-specific code from Instrumentation class into their own classes
