+++ src/Tests/Core/ImplementationDetails/Instrumentation/Instrumentation_of_properties_Tests.cs
@@ -3,7 +3,7 @@ using Fettle.Core.Internal.RoslynExtensions;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using NUnit.Framework;
 using System.Threading.Tasks;
-using InstrumentationImpl = Fettle.Core.Internal.Instrumentation;
+using InstrumentationImpl = Fettle.Core.Internal.Instrumentation.Instrumentation;
 
 namespace Fettle.Tests.Core.ImplementationDetails.Instrumentation
 {
